xway/requirements.txt
Django==3.2
gunicorn==20.1.0
djangorestframework==3.12.2
psycopg2-binary==2.8.5
requests